refactor: move test files under tests/
