+++ oxide_editor/src/models/tests/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+mod filesystem;
