+++ oxide_editor/src/viewmodels/tests/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+mod content_drawer;
